Main module completed with added basic structure of exec module
Made import statments more explicit to what the module was using instead of *. Added new command clear that clears the screen. Created the basic structure of the exec module with clear and quit implemented while creating the basic structure for add/set. Using the exec module completed the main module.